--- cov1.txt
@@ -1,32 +1,0 @@
-1.000 0.900 0.398 0.509 0.743 0.971 0.476 0.843 0.486 0.705 0.857 0.904 0.906 0.590 0.195 0.734 0.893 0.458 0.385 0.353 0.554 0.918 0.890 0.744 0.702 0.285 0.804 0.866 0.948 0.778 0.544 0.483
-0.900 1.000 0.642 0.764 0.866 0.961 0.500 0.937 0.730 0.875 0.879 0.800 0.794 0.552 0.221 0.653 0.762 0.432 0.289 0.454 0.499 0.820 0.779 0.699 0.708 0.474 0.871 0.839 0.967 0.916 0.693 0.670
-0.398 0.642 1.000 0.908 0.885 0.565 0.748 0.755 0.759 0.644 0.674 0.501 0.495 0.266 0.315 0.509 0.421 0.398 0.240 0.449 0.327 0.459 0.467 0.371 0.472 0.578 0.547 0.568 0.492 0.711 0.653 0.829
-0.509 0.764 0.908 1.000 0.864 0.660 0.633 0.878 0.940 0.857 0.774 0.594 0.587 0.530 0.528 0.627 0.540 0.568 0.384 0.684 0.495 0.590 0.589 0.574 0.662 0.756 0.707 0.662 0.650 0.808 0.771 0.909
-0.743 0.866 0.885 0.864 1.000 0.842 0.825 0.934 0.758 0.743 0.862 0.757 0.749 0.480 0.344 0.713 0.706 0.521 0.384 0.504 0.507 0.743 0.739 0.611 0.688 0.595 0.808 0.807 0.800 0.905 0.777 0.799
-0.971 0.961 0.565 0.660 0.842 1.000 0.544 0.915 0.597 0.792 0.899 0.897 0.887 0.561 0.199 0.707 0.861 0.461 0.367 0.398 0.578 0.909 0.879 0.766 0.750 0.400 0.868 0.916 0.969 0.876 0.653 0.627
-0.476 0.500 0.748 0.633 0.825 0.544 1.000 0.688 0.523 0.388 0.619 0.571 0.569 0.333 0.393 0.658 0.562 0.534 0.474 0.466 0.458 0.556 0.591 0.456 0.543 0.551 0.594 0.615 0.482 0.668 0.637 0.584
-0.843 0.937 0.755 0.878 0.934 0.915 0.688 1.000 0.844 0.888 0.934 0.852 0.850 0.681 0.473 0.802 0.816 0.619 0.476 0.631 0.619 0.855 0.845 0.759 0.807 0.661 0.896 0.879 0.906 0.928 0.809 0.813
-0.486 0.730 0.759 0.940 0.758 0.597 0.523 0.844 1.000 0.850 0.695 0.511 0.533 0.608 0.595 0.652 0.482 0.539 0.336 0.724 0.392 0.516 0.514 0.500 0.600 0.757 0.672 0.546 0.647 0.759 0.733 0.800
-0.705 0.875 0.644 0.857 0.743 0.792 0.388 0.888 0.850 1.000 0.864 0.736 0.738 0.653 0.470 0.666 0.718 0.615 0.448 0.699 0.592 0.765 0.734 0.735 0.726 0.624 0.782 0.743 0.808 0.801 0.660 0.722
-0.857 0.879 0.674 0.774 0.862 0.899 0.619 0.934 0.695 0.864 1.000 0.953 0.929 0.706 0.460 0.797 0.915 0.693 0.587 0.589 0.706 0.930 0.933 0.792 0.769 0.496 0.796 0.876 0.837 0.796 0.640 0.734
-0.904 0.800 0.501 0.594 0.757 0.897 0.571 0.852 0.511 0.736 0.953 1.000 0.980 0.707 0.415 0.812 0.973 0.665 0.611 0.488 0.734 0.976 0.980 0.825 0.776 0.381 0.757 0.905 0.814 0.695 0.560 0.625
-0.906 0.794 0.495 0.587 0.749 0.887 0.569 0.850 0.533 0.738 0.929 0.980 1.000 0.727 0.431 0.873 0.967 0.632 0.543 0.474 0.642 0.954 0.948 0.771 0.734 0.370 0.739 0.869 0.816 0.679 0.548 0.580
-0.590 0.552 0.266 0.530 0.480 0.561 0.333 0.681 0.608 0.653 0.706 0.707 0.727 1.000 0.817 0.823 0.749 0.750 0.620 0.679 0.639 0.711 0.713 0.692 0.720 0.595 0.645 0.592 0.613 0.504 0.590 0.569
-0.195 0.221 0.315 0.528 0.344 0.199 0.393 0.473 0.595 0.470 0.460 0.415 0.431 0.817 1.000 0.716 0.498 0.845 0.745 0.854 0.625 0.447 0.479 0.561 0.631 0.756 0.418 0.355 0.255 0.296 0.543 0.574
-0.734 0.653 0.509 0.627 0.713 0.707 0.658 0.802 0.652 0.666 0.797 0.812 0.873 0.823 0.716 1.000 0.871 0.767 0.629 0.686 0.594 0.820 0.823 0.692 0.715 0.564 0.676 0.702 0.703 0.625 0.601 0.585
-0.893 0.762 0.421 0.540 0.706 0.861 0.562 0.816 0.482 0.718 0.915 0.973 0.967 0.749 0.498 0.871 1.000 0.752 0.691 0.579 0.770 0.986 0.982 0.861 0.797 0.409 0.751 0.869 0.809 0.664 0.533 0.547
-0.458 0.432 0.398 0.568 0.521 0.461 0.534 0.619 0.539 0.615 0.693 0.665 0.632 0.750 0.845 0.767 0.752 1.000 0.948 0.912 0.877 0.727 0.759 0.801 0.774 0.675 0.573 0.593 0.466 0.490 0.523 0.575
-0.385 0.289 0.240 0.384 0.384 0.367 0.474 0.476 0.336 0.448 0.587 0.611 0.543 0.620 0.745 0.629 0.691 0.948 1.000 0.820 0.920 0.675 0.724 0.791 0.741 0.587 0.494 0.563 0.350 0.373 0.437 0.462
-0.353 0.454 0.449 0.684 0.504 0.398 0.466 0.631 0.724 0.699 0.589 0.488 0.474 0.679 0.854 0.686 0.579 0.912 0.820 1.000 0.758 0.583 0.598 0.732 0.750 0.831 0.599 0.510 0.466 0.551 0.615 0.618
-0.554 0.499 0.327 0.495 0.507 0.578 0.458 0.619 0.392 0.592 0.706 0.734 0.642 0.639 0.625 0.594 0.770 0.877 0.920 0.758 1.000 0.804 0.827 0.940 0.894 0.640 0.694 0.776 0.550 0.561 0.598 0.619
-0.918 0.820 0.459 0.590 0.743 0.909 0.556 0.855 0.516 0.765 0.930 0.976 0.954 0.711 0.447 0.820 0.986 0.727 0.675 0.583 0.804 1.000 0.991 0.908 0.851 0.458 0.820 0.928 0.861 0.742 0.602 0.607
-0.890 0.779 0.467 0.589 0.739 0.879 0.591 0.845 0.514 0.734 0.933 0.980 0.948 0.713 0.479 0.823 0.982 0.759 0.724 0.598 0.827 0.991 1.000 0.896 0.840 0.456 0.785 0.912 0.817 0.716 0.585 0.620
-0.744 0.699 0.371 0.574 0.611 0.766 0.456 0.759 0.500 0.735 0.792 0.825 0.771 0.692 0.561 0.692 0.861 0.801 0.791 0.732 0.940 0.908 0.896 1.000 0.962 0.662 0.845 0.899 0.761 0.710 0.697 0.651
-0.702 0.708 0.472 0.662 0.688 0.750 0.543 0.807 0.600 0.726 0.769 0.776 0.734 0.720 0.631 0.715 0.797 0.774 0.741 0.750 0.894 0.851 0.840 0.962 1.000 0.804 0.908 0.908 0.764 0.779 0.855 0.765
-0.285 0.474 0.578 0.756 0.595 0.400 0.551 0.661 0.757 0.624 0.496 0.381 0.370 0.595 0.756 0.564 0.409 0.675 0.587 0.831 0.640 0.458 0.456 0.662 0.804 1.000 0.737 0.586 0.476 0.676 0.897 0.782
-0.804 0.871 0.547 0.707 0.808 0.868 0.594 0.896 0.672 0.782 0.796 0.757 0.739 0.645 0.418 0.676 0.751 0.573 0.494 0.599 0.694 0.820 0.785 0.845 0.908 0.737 1.000 0.913 0.908 0.939 0.902 0.708
-0.866 0.839 0.568 0.662 0.807 0.916 0.615 0.879 0.546 0.743 0.876 0.905 0.869 0.592 0.355 0.702 0.869 0.593 0.563 0.510 0.776 0.928 0.912 0.899 0.908 0.586 0.913 1.000 0.862 0.840 0.782 0.732
-0.948 0.967 0.492 0.650 0.800 0.969 0.482 0.906 0.647 0.808 0.837 0.814 0.816 0.613 0.255 0.703 0.809 0.466 0.350 0.466 0.550 0.861 0.817 0.761 0.764 0.476 0.908 0.862 1.000 0.904 0.698 0.587
-0.778 0.916 0.711 0.808 0.905 0.876 0.668 0.928 0.759 0.801 0.796 0.695 0.679 0.504 0.296 0.625 0.664 0.490 0.373 0.551 0.561 0.742 0.716 0.710 0.779 0.676 0.939 0.840 0.904 1.000 0.852 0.732
-0.544 0.693 0.653 0.771 0.777 0.653 0.637 0.809 0.733 0.660 0.640 0.560 0.548 0.590 0.543 0.601 0.533 0.523 0.437 0.615 0.598 0.602 0.585 0.697 0.855 0.897 0.902 0.782 0.698 0.852 1.000 0.833
-0.483 0.670 0.829 0.909 0.799 0.627 0.584 0.813 0.800 0.722 0.734 0.625 0.580 0.569 0.574 0.585 0.547 0.575 0.462 0.618 0.619 0.607 0.620 0.651 0.765 0.782 0.708 0.732 0.587 0.732 0.833 1.000
