--- schedule.bat
@@ -1,8 +1,0 @@
-python train.py --windowSize 128 --modelName CNN --bandL 0.1 --bandR 50 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN --bandL 0.1 --bandR 8 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN --bandL 8 --bandR 13 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN_LSTM --bandL 0.1 --bandR 50 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN_LSTM --bandL 0.1 --bandR 8 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN_LSTM --bandL 8 --bandR 13 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-python train.py --windowSize 128 --modelName CNN_LSTM --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA
