--- schedule2.bat
@@ -1,8 +1,0 @@
-python train.py --windowSize 128 --modelName PSD --output demoDEA.txt --bandL 0.1 --bandR 50 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName PSD --output demoDEA.txt --bandL 0.1 --bandR 8 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName PSD --output demoDEA.txt --bandL 8 --bandR 13 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName PSD --output demoDEA.txt --bandL 13 --bandR 30 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName IHAR --output demoDEA.txt --bandL 0.1 --bandR 50 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName IHAR --output demoDEA.txt --bandL 0.1 --bandR 8 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName IHAR --output demoDEA.txt --bandL 8 --bandR 13 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py --windowSize 128 --modelName IHAR --output demoDEA.txt --bandL 13 --bandR 30 --extractFixation False --thinking False --trainTestSeperate False --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA
