--- schedule4.bat
@@ -1,15 +1,0 @@
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm EA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 128 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 64 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 64 --input C:\Users\hmi\Desktop\DataVIN --eaNorm EA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 64 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 256 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 256 --input C:\Users\hmi\Desktop\DataVIN --eaNorm EA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 256 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA 
-
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 192 --input C:\Users\hmi\Desktop\DataVIN --eaNorm DEA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 192 --input C:\Users\hmi\Desktop\DataVIN --eaNorm EA 
-python train.py  --modelName SHALLOW --output duration.txt --bandL 13 --bandR 30 --extractFixation False --thinking True --trainTestSeperate True --trainTestSession False --windowSize 192 --input C:\Users\hmi\Desktop\DataVIN --eaNorm 1EA
