+++ util/preproc.py
@@ -14,6 +14,7 @@ import sys
 from scipy import signal
 import time
 from sklearn.utils import shuffle
+from scipy.linalg import sqrtm
 
 def balance_set(X,y,replace=False):
     labels= np.unique(y)
@@ -35,10 +36,12 @@ def balance_set(X,y,replace=False):
             balanced_data = np.concatenate((balanced_data,tempx))
             balanced_label = np.concatenate((balanced_label,tempy))
     return shuffle(balanced_data,balanced_label)
+
+
 def plot_confusion_matrix(y_true, y_pred, classes,
-                          normalize=False,
-                          title=None,
-                          cmap=plt.cm.Blues):
+                            normalize=False,
+                            title=None,
+                            cmap=plt.cm.Blues):
     """
     This function prints and plots the confusion matrix.
     Normalization can be applied by setting `normalize=True`.
@@ -60,34 +63,35 @@ def plot_confusion_matrix(y_true, y_pred, classes,
     else:
         print('Confusion matrix, without normalization')
 
-    
 
-    fig, ax = plt.subplots()
-    im = ax.imshow(cm, interpolation='nearest', cmap=cmap)
-    ax.figure.colorbar(im, ax=ax)
-    # We want to show all ticks...
-    ax.set(xticks=np.arange(cm.shape[1]),
-           yticks=np.arange(cm.shape[0]),
-           # ... and label them with the respective list entries
-           xticklabels=classes, yticklabels=classes,
-           title=title,
-           ylabel='True label',
-           xlabel='Predicted label')
 
-    # Rotate the tick labels and set their alignment.
-    plt.setp(ax.get_xticklabels(), rotation=45, ha="right",
-             rotation_mode="anchor")
+    # fig, ax = plt.subplots()
+    # im = ax.imshow(cm, interpolation='nearest', cmap=cmap)
+    # ax.figure.colorbar(im, ax=ax)
+    # # We want to show all ticks...
+    # ax.set(xticks=np.arange(cm.shape[1]),
+    #     yticks=np.arange(cm.shape[0]),
+    #     # ... and label them with the respective list entries
+    #     xticklabels=classes, yticklabels=classes,
+    #     title=title,
+    #     ylabel='True label',
+    #     xlabel='Predicted label')
 
-    # Loop over data dimensions and create text annotations.
-    fmt = '.2f' if normalize else 'd'
-    thresh = cm.max() / 2.
-    for i in range(cm.shape[0]):
-        for j in range(cm.shape[1]):
-            ax.text(j, i, format(cm[i, j], fmt),
-                    ha="center", va="center",
-                    color="white" if cm[i, j] > thresh else "black")
-    fig.tight_layout()
-    return ax
+    # # Rotate the tick labels and set their alignment.
+    # plt.setp(ax.get_xticklabels(), rotation=45, ha="right",
+    #     rotation_mode="anchor")
+
+    # # Loop over data dimensions and create text annotations.
+    # fmt = '.2f' if normalize else 'd'
+    # thresh = cm.max() / 2.
+    # for i in range(cm.shape[0]):
+    #     for j in range(cm.shape[1]):
+    #         ax.text(j, i, format(cm[i, j], fmt),
+    #                 ha="center", va="center",
+    #                 color="white" if cm[i, j] > thresh else "black")
+    # fig.tight_layout()
+    # plt.show()
+    # return ax
 
 def butter_bandpass_filter(data, lowcut, highcut, fs, order):
     nyq = 0.5 * fs
@@ -130,3 +134,202 @@ def filter_notch(data,notchcut,fs):
 #     data = signal.filtfilt(b, a, data)
     data= signal.lfilter(i, u, data)
     return(data)
+
+def calculateAdj(inputMat):
+    mat = np.asarray(inputMat)
+    s1, s2, s3, s4 = mat.shape
+    mat = mat.reshape(s1 * s2, s3, s4)
+    coMat = mat.mean(axis=0)
+    Adj = np.corrcoef(coMat[:, :].T)
+    # print(Adj)
+    D = np.array(np.sum(Adj, axis=0))
+    D = np.sqrt(D)
+    # print(np.matrix(np.diag(D)).shape)
+    # print(np.diag(D).shape)
+    # print(np.diag(D))
+    # print(np.matrix(np.diag(D)))
+
+    # print(np.diag(D).shape)
+    # print(np.matrix(np.diag(D)).shape)
+
+    D = np.diag(D)
+    matAdj = torch.Tensor(np.linalg.inv(D) * Adj * np.linalg.inv(D)).to(device)
+    # matAdj = get_adj(Adj)
+    return matAdj
+
+def get_adj(adj):
+    """
+    build symmetric adjacency matrix
+    @param adj:
+    @return:
+    """
+    adj = adj + adj.T.multiply(adj.T > adj) - adj.multiply(adj.T > adj)
+    adj = normalize(adj + sp.eye(adj.shape[0]))
+    adj = sparse_mx_to_torch_sparse_tensor(adj)
+    return adj
+
+
+def normalize(mx):
+    """Row-normalize sparse matrix"""
+    rowsum = np.array(mx.sum(1))
+    r_inv = np.power(rowsum, -1).flatten()
+    r_inv[np.isinf(r_inv)] = 0.
+    r_mat_inv = sp.diags(r_inv)
+    mx = r_mat_inv.dot(mx)
+    return mx
+
+def getNormR(data, shapeIn = 32):
+    print(data.shape)
+    normR = np.zeros([shapeIn, shapeIn])
+    for ii in range(len(data)):
+        covX = np.matmul(data[ii].T, data[ii])
+        normR += covX
+    normR = normR / len(data)
+    normR = sqrtm(normR)
+    normR = np.linalg.inv(normR)
+    return normR
+
+
+# Calculate FFT of data Pos or Neg
+def GetFFT(datas):
+    sampling_length = 128
+    fft_rs = []
+    for data in datas:
+        ts = 1.0/ 128
+
+        freq = np.fft.fftfreq(len(data), d = ts)
+        fft = np.fft.fft(data)
+        fft_rs.append(fft)
+
+    return fft_rs, freq
+
+# Calculate Power Spectral Density(PSD) (equation 1)
+def GetPSD(datas):
+    sampling_length = 128
+    PSD_rs = []
+    for data in datas:
+        PSD_seg = []
+        for d in data:
+            PSD_n = d * np.conjugate(d)/128
+            PSD_seg.append(PSD_n)
+        PSD_rs.append(PSD_seg)
+  
+    return PSD_rs
+
+def moving_average(x, w):
+    return np.convolve(x, np.ones(w), 'valid') / w
+
+def MA(datas, windowSize = 4):
+    newData = []
+    for data in datas:
+        tmp = moving_average(data, windowSize)
+        newData.append(tmp)
+
+    return np.vstack(newData)
+
+
+def matmul_list(matrix_list, db=False):
+    if db:
+        for x in matrix_list:
+            print(x.shape)
+    number_matrix = len(matrix_list)
+    result = np.copy(matrix_list[0])
+    for i in range(1, number_matrix):
+        result = np.matmul(result, matrix_list[i])
+    return result
+
+def setting_rank(eigen_vector):
+    minCumSV = 0.9
+    current_sum = 0
+    sum_list = np.sum(eigen_vector)
+    for x in range(len(eigen_vector)):
+        current_sum += eigen_vector[x]
+        if current_sum > minCumSV * sum_list:
+            return x + 1
+    return len(eigen_vector)
+
+
+def applyNorm(X_train, normMat):
+    tmp = []
+    for ii in range(len(X_train)):
+        Xnew = np.matmul(X_train[ii], normMat)
+        tmp.append(Xnew)
+    return np.asarray(tmp)
+
+
+def setSeed(seed):
+    random.seed(seed)
+    os.environ['PYTHONHASHSEED'] = str(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+
+def PSD(X_train, y_train, X_test, y_test):
+    tmp = []
+    for xxx in X_train:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_train = np.vstack(tmp)
+
+    tmp = []
+    for xxx in X_test:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_test = np.vstack(tmp)
+    return SVM(X_train, y_train, X_test, y_test)
+
+
+def IHAR(X_train, y_train, X_test, y_test, listChns):
+    
+    electrodeIHAR = [['Fp1', 'F7', 'F3', 'FC5', 'T7', 'P7', 'O1'], ['Fp2', 'F8', 'F4', 'FC6', 'T8', 'P8', 'O2']]
+    numNode = len(electrodeIHAR[0])
+    electrodeIndex = []
+    for electrodes in electrodeIHAR:
+        electrodeIndex.append([listChns.index(x) for x in electrodes])
+    tmp = []
+    for xxx in X_train:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = MA(newX, args.windowIHAR)
+        IharTrain = []
+        for ii in range(numNode):
+            left = newX[electrodeIndex[0][ii]]
+            right = newX[electrodeIndex[1][ii]]
+            ihar = left / right
+            IharTrain.append(ihar)
+        newX = np.vstack([np.vstack(IharTrain), newX])
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_train = np.vstack(tmp)
+
+    tmp = []
+    for xxx in X_test:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = MA(newX, args.windowIHAR)
+        IharTrain = []
+        for ii in range(numNode):
+            left = newX[electrodeIndex[0][ii]]
+            right = newX[electrodeIndex[1][ii]]
+            ihar = left / right
+            IharTrain.append(ihar)
+        newX = np.vstack([np.vstack(IharTrain), newX])
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_test = np.vstack(tmp)    
+    return SVM(X_train, y_train, X_test, y_test)
